+++ lib/improver/utilities/cube_manipulation.py
@@ -39,7 +39,6 @@ import iris
 from iris.coords import AuxCoord, DimCoord
 from iris.exceptions import CoordinateNotFoundError
 
-from improver.blending.weighted_blend import rationalise_blend_time_coords
 from improver.utilities.cube_checker import (
     check_cube_coordinates, check_cube_not_float64, find_threshold_coordinate)
 
@@ -194,6 +193,7 @@ class ConcatenateCubes():
                     cube.add_aux_coord(
                         temp_aux_coord,
                         data_dims=coord_names.index(self.master_coord))
+      
         return cube
 
     @staticmethod
@@ -599,149 +599,6 @@ def merge_cubes(cubes):
     """
     result = MergeCubes().process(cubes)
     return result
-
-
-# TODO move this class into improver.blending.weighted_blend
-class MergeCubesForWeightedBlending():
-    """Prepares cubes for cycle and grid blending"""
-
-    def __init__(self, blend_coord, weighting_coord=None, model_id_attr=None):
-        """
-        Initialise the class
-
-        Args:
-            blend_coord (str):
-                Name of coordinate over which blending will be performed.  For
-                multi-model blending this is flexible to any string containing
-                "model".  For all other coordinates this is prescriptive:
-                cube.coord(blend_coord) must return an iris.coords.Coord
-                instance for all cubes passed into the "process" method.
-
-        Kwargs:
-            weighting_coord (str or None):
-                The coordinate across which weights will be scaled in a
-                multi-model blend.  Required for
-                rationalise_blend_time_coordinates.
-            model_id_attr (str or None):
-                Name of attribute used to identify model for grid blending.
-                None for cycle blending.
-
-        Raises:
-            ValueError:
-                If trying to blend over model when model_id_attr is not set
-        """
-        if "model" in blend_coord and model_id_attr is None:
-            raise ValueError(
-                "model_id_attr required to blend over {}".format(blend_coord))
-
-        # ensure model coordinates are not created for non-model blending
-        if "model" not in blend_coord and model_id_attr is not None:
-            warnings.warn(
-                "model_id_attr not required for blending over {} - "
-                "will be ignored".format(blend_coord))
-            model_id_attr = None
-
-        self.blend_coord = blend_coord
-        self.weighting_coord = weighting_coord
-        self.model_id_attr = model_id_attr
-
-    def _create_model_coordinates(self, cubelist):
-        """
-        Adds numerical model ID and string model configuration scalar
-        coordinates to input cubes if self.model_id_attr is specified.
-        Sets the original attribute value to "blend", in anticipation.
-        Modifies cubes in place.
-
-        Args:
-            cubelist (iris.cube.CubeList):
-                List of cubes to be merged for blending
-
-        Raises:
-            ValueError:
-                If self.model_id_attr is not present on all cubes
-            ValueError:
-                If input cubelist contains cubes from the same model
-        """
-        model_titles = []
-        for i, cube in enumerate(cubelist):
-            if self.model_id_attr not in cube.attributes:
-                msg = ('Cannot create model ID coordinate for grid blending '
-                       'as "model_id_attr={}" was not found within the cube '
-                       'attributes'.format(self.model_id_attr))
-                raise ValueError(msg)
-
-            model_title = cube.attributes.pop(self.model_id_attr)
-            if model_title in model_titles:
-                raise ValueError('Cannot create model dimension coordinate '
-                                 'with duplicate points')
-            model_titles.append(model_title)
-            cube.attributes[self.model_id_attr] = "blend"
-
-            new_model_id_coord = build_coordinate([1000 * i],
-                                                  long_name='model_id',
-                                                  data_type=np.int32)
-            new_model_coord = (
-                build_coordinate([model_title],
-                                 long_name='model_configuration',
-                                 coord_type=AuxCoord,
-                                 data_type=np.str))
-
-            cube.add_aux_coord(new_model_id_coord)
-            cube.add_aux_coord(new_model_coord)
-
-    def process(self, cubes_in, cycletime=None):
-        """
-        Prepares merged input cube for cycle and grid blending
-
-        Args:
-            cubes (iris.cube.CubeList or iris.cube.Cube):
-                Cubes to be merged.
-
-        Kwargs:
-            cycletime (str or None):
-                The cycletime in a YYYYMMDDTHHMMZ format e.g. 20171122T0100Z.
-                Can be used in rationalise_blend_time_coordinates.
-
-        Returns:
-            iris.cube.Cube:
-                Merged cube.
-
-        Raises:
-            ValueError:
-                If self.blend_coord is not present on all cubes (unless
-                blending over models)
-        """
-        # if input is already a single cube, return unchanged
-        if isinstance(cubes_in, iris.cube.Cube):
-            return cubes_in
-
-        if len(cubes_in) == 1:
-            return cubes_in[0]
-
-        # create copies of input cubes so as not to modify in place
-        cubelist = iris.cube.CubeList([])
-        for cube in cubes_in:
-            if ("model" not in self.blend_coord and
-                    not cube.coords(self.blend_coord)):
-                raise ValueError(
-                    "{} coordinate is not present on all input "
-                    "cubes".format(self.blend_coord))
-            cubelist.append(cube.copy())
-
-        # TODO move rationalise_blend_time_coords into this class
-        rationalise_blend_time_coords(
-            cubelist, self.blend_coord, cycletime=cycletime,
-            weighting_coord=self.weighting_coord)
-
-        # create model ID and model configuration coordinates if blending
-        # different models
-        if self.model_id_attr is not None:
-            self._create_model_coordinates(cubelist)
-
-        # merge resulting cubelist
-        result = MergeCubes().process(cubelist, check_time_bounds_ranges=True)
-
-        return result
 
 
 def get_filtered_attributes(cube, attribute_filter=None):
